+++ lib/rubocop/cop/vicenzo_cops.rb
@@ -1,1 +1,3 @@
 # frozen_string_literal: true
+
+require_relative 'vicenzo/rspec/nested_let_redefinition'