+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## [0.1.0] - 2025-03-21
 
-- Initial release
+- Initial release;
+- Add RoboCop::Cop::Vicenzo::RSpec::NestedLetRedefinition #1;